Create distinct LBDatabase and LBManager, remove batsyncer, various cleanup
