+++ src/ck-cp/controlPoints.h
@@ -37,7 +37,7 @@
 #include <float.h>
 #include <charm-api.h>
 
-#include "LBDatabase.h"
+#include "LBManager.h"
 #include "arrayRedistributor.h"
 #include "pathHistory.h"
