+++ src/ck-cp/pathHistory.C
@@ -9,7 +9,7 @@
 
 //#define DEBUG  1
 #include "PathHistory.decl.h"
-#include "LBDatabase.h"
+#include "LBManager.h"
 #include "pathHistory.h"
 #include <register.h>
