+++ src/ck-ldb/CentralLB.h
@@ -99,8 +99,9 @@ public:
   int GetPESpeed();
   inline void setConcurrent(bool c) { concurrent = c; }
 
-  static void staticAtSync(void*);
-  void AtSync(void); // Everything is at the PE barrier
+//  static void staticAtSync(void*);
+//  void AtSync(void); // Everything is at the PE barrier
+  void InvokeLB();
   void ProcessAtSync(void); // Receive a message from AtSync to avoid
                             // making projections output look funny
   void SendStats();
