+++ src/ck-ldb/DistBaseLB.ci
@@ -1,6 +1,6 @@
 module DistBaseLB {
 
-extern module LBDatabase;
+extern module LBManager;
 extern module BaseLB;
 
 
@@ -8,7 +8,7 @@ group [migratable] DistBaseLB: BaseLB {
   entry void DistBaseLB(const CkLBOptions &);  
   entry void LoadBalance();
   entry [reductiontarget] void barrierDone();
-	entry void AtSync();
+	entry void InvokeLB();
   entry [reductiontarget] void ResumeClients();
   entry void ResumeClients(int balancing);
 };
