+++ src/ck-ldb/DistributedLB.C
@@ -34,11 +34,9 @@ void DistributedLB::initnodeFn()
 void DistributedLB::turnOn()
 {
 #if CMK_LBDB_ON
-  theLbdb->
+  lbmgr->
     TurnOnBarrierReceiver(receiver);
-  theLbdb->
-    TurnOnNotifyMigrated(notifier);
-  theLbdb->
+  lbmgr->
     TurnOnStartLBFn(startLbFnHdl);
 #endif
 }
@@ -46,11 +44,9 @@ void DistributedLB::turnOn()
 void DistributedLB::turnOff()
 {
 #if CMK_LBDB_ON
-  theLbdb->
+  lbmgr->
     TurnOffBarrierReceiver(receiver);
-  theLbdb->
-    TurnOffNotifyMigrated(notifier);
-  theLbdb->
+  lbmgr->
     TurnOffStartLBFn(startLbFnHdl);
 #endif
 }
@@ -362,7 +358,7 @@ void DistributedLB::AfterLBReduction(CkReductionMsg* redn_msg) {
     Cleanup();
     PackAndSendMigrateMsgs();
     if (!(_lb_args.metaLbOn() && _lb_args.metaLbModelDir() != nullptr))
-      theLbdb->nextLoadbalancer(seqno);
+      lbmgr->nextLoadbalancer(seqno);
   }
   delete [] results;
 }
