+++ src/ck-ldb/HybridBaseLB.ci
@@ -1,6 +1,6 @@
 module HybridBaseLB {
 
-extern module LBDatabase;
+extern module LBManager;
 extern module BaseLB;
 
 include "CentralLBMsg.h";
