--- src/ck-ldb/LBDatabase.ci
@@ -1,17 +1,0 @@
-module LBDatabase {
-  extern module CkMarshall;
-  extern module NullLB;
-
-  readonly CkGroupID _lbdb;
-
-  mainchare LBDBInit {
-    entry LBDBInit(CkArgMsg *m);
-  };
-
-  group [migratable] LBDatabase {
-    entry void LBDatabase(void);  
-    entry void ResumeClients();
-    initnode void initnodeFn();
-  };
-
-};
