--- src/ck-ldb/NullLB.ci
@@ -1,14 +1,0 @@
-module NullLB {
-
-extern module LBDatabase;
-extern module BaseLB;
-
-initnode void lbinit(void);
-initproc void lbprocinit(void);
-
-group [migratable] NullLB: BaseLB {
-  entry void NullLB(const CkLBOptions &);  
-  entry void migrationsDone(void);
-};
-
-};
