+++ src/ck-ldb/lbdb.h
@@ -19,7 +19,7 @@
 #include <inttypes.h>
 #include <list>
 
-class LBDatabase;//Forward declaration
+class LBManager;//Forward declaration
 
 //typedef float floatType;
 // type defined by build option
@@ -252,7 +252,7 @@ typedef struct _LDCommData {
 /*
  * Requests to load balancer
  *   FIXME: these routines don't seem to exist anywhere-- are they obsolete?
- *   Are the official versions now in LBDatabase.h?
+ *   Are the official versions now in LBManager.h?
  */
 void LBBalance(void *param);
 void LBCollectStatsOn(void);
@@ -295,12 +295,6 @@ typedef void (*LDBarrierFn)(void *user_ptr);
 typedef void (*LDResumeFn)(void *user_ptr);
 
 class client;
-struct LDBarrierClient {
-  std::list<client *>::iterator i;
-  LDBarrierClient() { }
-  LDBarrierClient(std::list<client *>::iterator in)
-  : i(in) { }
-};
 
 class receiver;
 struct LDBarrierReceiver {
