+++ src/conv-ldb/topology.C
@@ -5,7 +5,7 @@
 
 #include <math.h>
 
-//#include "LBDatabase.h"
+//#include "LBManager.h"
 #include <vector>
 #include "converse.h"
 #include "topology.h"
